logic for posting ai prompts done
