+++ app/create-prompt/page.jsx
@@ -1,0 +1,43 @@
+"use client";
+
+import Form from "@components/Form";
+import { useSession } from "next-auth/react";
+import { useRouter } from "next/navigation";
+import { Router } from "next/router";
+import React, { useState } from "react";
+
+const CreatePrompt = () => {
+  const [submitting, setSubmitting] = useState(false);
+  const [post, setPost] = useState({ prompt: "", tag: "" });
+
+  const createPrompt = async (e) => {
+    e.preventDefault();
+    setSubmitting(true);
+
+    try {
+      //post whatever is in the text post to the api we's gonna create called "prompt from inside the api folder"
+      const res = await fetch("/api/prompt/new", {
+        method: "POST",
+        body: JSON.stringify({
+          prompt: post.prompt,
+          userId: session?.user.id,
+          tag: post.tag,
+        }),
+      });
+      if (res.ok) {
+        Router.push("/");
+      }
+    } catch (error) {
+      console.log(error);
+    } finally {
+      setSubmitting(false);
+    }
+  };
+  return (
+    <div>
+      <Form type="Create" post={post} setPost={setPost} submitting={submitting} handleSubmit={createPrompt} />
+    </div>
+  );
+};
+
+export default CreatePrompt;
